+++ CHANGELOG.md
@@ -12,6 +12,11 @@
 
 1. [#127](https://github.com/InfluxCommunity/influxdb3-go/pull/127): LPBatcher now returns first line of the internal buffer when the line length exceeds the batch size.
 
+### Features
+
+1. [#131](https://github.com/InfluxCommunity/influxdb3-go/pull/131): Add new PointValueIterator based on google
+   guidelines [Guidelines](https://github.com/googleapis/google-cloud-go/wiki/Iterator-Guidelines)
+
 ## 2.0.0 [2024-12-13]
 
 ### Breaking Changes